moving dispatch to domEvents
